DifferentialEquations.jl compatibility
Added several functions to create an interface with DifferentialEquations.jl. The main idea is to make ParticleField be treated like a 2D array and then write broadcasting functions to make that the expected array behavior is still valid. The old functionality still works, but now 2D array indexing is supported (as well as any operations that make sense on a 2D matrix). These operations allow DifferentialEquations.jl to use ParticleField objects when solving systems of ODEs.

A second small change was made to allow passing a different base float type to ParticleField. The default is still the FMM's Real type.