DifferentialEquaitons.jl compatibility
Adds an interface so particles can be handled by DifferentialEquations.jl. They are treated like 1D arrays with 9 elements (3 position elements, 3 vorticity elements, and 1 each for sigma, vol, and circulation). A small change was also made to allow for other floating point types to be passed to the particle constructor.