DifferentialEquations.jl compatibility
Current state of the file that actually runs simulations. While simulations can be run with the full VPM (with FMM + viscous diffusion if desired), there is still some functionality that is not added yet. The system state is not saved at intermediate steps for the fullstep method (yet) and will require another callback. Also, options like runtime functions do not work yet. Eventually, everything should be able to be set from the original VPM interface but run DifferentialEquations.jl under the hood.