Saving no longer breaks with AD
Also fixed a bug where all particles were saved in the same memory location; I think this happened anyway but the FMM re-allocated the memory afterwards. The solution was to explicitly write Base.zeros(Particle) and make sure that a full array of particles was being allocated.

ReverseAD is getting closer. It might be working already (since I rewrote the part of the particle interaction function with the square root evaluations at zero), but the memory usage is quite large and I'll need to test ReverseAD with really small simulations.