Progress towards ImplicitAD integration
Got nonzero derivatives from ImplicitAD with forward AD. The issue was that on the first step of the second pass I needed to explicitly return the initial conditions to ensure that the associated derivative values were used. In other words, in the stepping function that is passed to ImplicitAD, I had to have check if t == tprev (which only occurs at the very beginning) and if so just return the input. Reverse AD still returns zeros (as well as having intermediate derivative values of zero) and the forward AD values are wrong, so I suspect that I have a similar issue at the end of the solution process.
I also removed a bunch of code I wrote but didn't end up using because of ImplicitAD.
Finally, I added a setting "mode" to the run_vpm! function to toggle ImplicitAD. This lets me check to see if my derivative values are correct (when compared with pure AD without ImplicitAD).