Compatiblity with new particle instantiation/runtime functions, part 1
New particle instantiation is compatible with ForwardDiff + ImplicitAD. It runs with ReverseDiff + ImplicitAD but returns zeros; I think the issue is that the computational graph is not correctly recorded. Both AD methods work correctly when use_implicitAD is false. This all assumes that the runtime function used to generate new particles is AD compatible. I also cleaned up various sections of code.